--- a/schema/spacewalk/susemanager-schema.changes.oholecek.optimize-system-overview
+++ b/schema/spacewalk/susemanager-schema.changes.oholecek.optimize-system-overview
@@ -1,0 +1,3 @@
+- Optimize system overview table generation
+  * Added new indexes
+  * Simplified queries