## What does this PR change?

Optimize system overview table generation:
- added two missing indexes to prevent seq scans
- simplified two queries:
  - computing of various erratas refactored from 3 queries to one. From ~2.7ms down to ~0.5ms
  - changed extra pkg count from massive outer join to anti join. From ~20ms down to ~10ms

Currently one iteration for one server in my small system ( 14 servers, synced SL Micro 5.5, 6.0, 6.1; SLE 15 SP5, SP6, SLE 12 SP5 channels):
```
duration: 0.057 ms
duration: 0.087 ms
duration: 0.028 ms
duration: 0.021 ms
duration: 0.482 ms
duration: 0.525 ms
duration: 0.012 ms
duration: 0.027 ms
duration: 0.021 ms
duration: 0.022 ms
duration: 0.015 ms
duration: 0.043 ms
duration: 0.009 ms
duration: 0.003 ms
duration: 0.125 ms
duration: 0.069 ms
duration: 10.328 ms
duration: 0.045 ms
duration: 0.009 ms
duration: 0.021 ms
duration: 0.020 ms
duration: 0.158 ms
duration: 0.058 ms
-- reported total for call:
duration: 27.605 ms
```

Original:
```
duration: 0.038 ms
duration: 0.034 ms
duration: 0.035 ms
duration: 0.048 ms
duration: 0.881 ms
duration: 0.874 ms
duration: 0.826 ms
duration: 1.039 ms
duration: 0.010 ms
duration: 0.026 ms
duration: 0.022 ms
duration: 0.025 ms
duration: 0.017 ms
duration: 0.021 ms
duration: 0.006 ms
duration: 0.002 ms
duration: 0.042 ms
duration: 0.038 ms
duration: 19.905 ms
duration: 1.452 ms
duration: 0.010 ms
duration: 0.025 ms
duration: 0.021 ms
duration: 0.414 ms
duration: 0.058 ms
-- reported total for call:
duration: 42.068 ms
```

Longest time (~27ms) is spent on a write.

## GUI diff

No difference.

- [X] **DONE**

## Documentation
- No documentation needed: only internal and user invisible changes

- [X] **DONE**

## Test coverage
ℹ️ If a major new functionality is added, it is **strongly recommended** that tests for the new functionality are added to the Cucumber test suite
- No tests: **add explanation**
- No tests: already covered
- Unit tests were added
- Cucumber tests were added

- [ ] **DONE**

## Links

Issue(s): #
Port(s): # **add downstream PR(s), if any**

- [ ] **DONE**

## Changelogs

Make sure the changelogs entries you are adding are compliant with https://github.com/uyuni-project/uyuni/wiki/Contributing#changelogs and https://github.com/uyuni-project/uyuni/wiki/Contributing#uyuni-projectuyuni-repository

If you don't need a changelog check, please mark this checkbox:

- [ ] No changelog needed

If you uncheck the checkbox after the PR is created, you will need to re-run `changelog_test` (see below)

## Re-run a test

If you need to re-run a test, please mark the related checkbox, it will be unchecked automatically once it has re-run:

- [ ] Re-run test "changelog_test"
- [ ] Re-run test "backend_unittests_pgsql"
- [ ] Re-run test "java_pgsql_tests"
- [ ] Re-run test "schema_migration_test_pgsql"
- [ ] Re-run test "susemanager_unittests"
- [ ] Re-run test "javascript_lint"
- [ ] Re-run test "spacecmd_unittests"

# Before you merge

Check [How to branch and merge properly](https://github.com/uyuni-project/uyuni/wiki/How-to-branch-and-merge-properly)!